// src/types/events.ts
export interface Event {
  id: number;
  idcategory: number;
  category: string;
  price: number;
  date: string;
  time: string;
  title: string;
  description: string;
  image: string;
  currency: string;
  addressimage: string;
  addressurl: string;
  place: string;
  obs: string;
  speaker: string;
  speaker_photo?: string;
  status: number;
} 